requirements.txt
python-telegram-bot==13.7
requests==2.26.0
yfinance==0.2.18
alpha_vantage==2.3.1
newsapi-python==0.2.6
beautifulsoup4==4.10.0
html5lib==1.1
textblob==0.15.3
pandas==1.3.4
schedule==1.1.0